// frontend/components/Modals/Analyze/Classify/k-medoids-cluster/types/output.ts
export interface ElbowPoint {
    k: number;
    totalCost: number;
    silhouetteScore: number;
}